eliminar imagenes
se eliminan las imagenes guardas en la base de datos y la carpeta ecommerce

+++ ecommerce/src/main/java/com/ecommerce/controller/ProductoController.java
@@ -2,6 +2,7 @@ package com.ecommerce.controller;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -19,6 +20,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.ecommerce.model.Imagen;
 import com.ecommerce.model.Producto;
+import com.ecommerce.model.Proveedor;
 import com.ecommerce.service.ICategoriaService;
 import com.ecommerce.service.IProductoService;
 import com.ecommerce.service.IProveedorService;
@@ -75,13 +77,80 @@ public class ProductoController {
 		return "redirect:/producto";  
 	}
 	
+	@GetMapping("/editar/{idProducto}")
+	public String editar(@PathVariable Integer idProducto, Model model) {
+	    Optional<Producto> optionalProducto = productoService.get(idProducto);
+	    if (optionalProducto.isPresent()) {
+	        Producto producto = optionalProducto.get();
+	        List<Imagen> imagenes = producto.getImagenesProducto();
+	        model.addAttribute("producto", producto);
+	        model.addAttribute("imagenes", imagenes);
+	        model.addAttribute("categoria", categoriaService.getAllCategoria());
+	        model.addAttribute("proveedor", proveedorService.getAllProveedor());
+	        return "producto/editar";
+	    } else {
+	        // Manejar el caso en el que el producto no existe
+	        return "redirect:/producto";
+	    }
+	}
+
+
+
 	
+	@PostMapping("/actualizar")
+	public String update(Producto producto, @RequestParam("nuevasImagenes") List<MultipartFile> nuevasImagenes, RedirectAttributes attribute){
+		
+		// Procesar las nuevas imágenes
+	    if (!nuevasImagenes.isEmpty()) {
+	        // Ruta absoluta donde se guardarán las imágenes
+	        String rutaAbsoluta = "D://Ecommerce//imagenesProducto";
+
+	        for (MultipartFile nuevaImagen : nuevasImagenes) {
+	            try {
+	                byte[] bytesImagen = nuevaImagen.getBytes();
+	                Path rutaCompleta = Paths.get(rutaAbsoluta + "//" + nuevaImagen.getOriginalFilename());
+	                Files.write(rutaCompleta, bytesImagen);
+	                Imagen nuevaImagenProducto = new Imagen();
+	                nuevaImagenProducto.setArchivoImagen(nuevaImagen.getOriginalFilename());
+	                nuevaImagenProducto.setProducto(producto);
+	                producto.getImagenesProducto().add(nuevaImagenProducto);
+	            } catch (IOException e) {
+	                e.printStackTrace();
+	            }
+	        }
+	    }
+		
+		productoService.update(producto);
+		attribute.addFlashAttribute("editar", "Datos actualizados con éxito!");
+		return "redirect:/producto";
+	}
 	
 	@GetMapping("/eliminar/{idProducto}")
 	public String eliminar(@PathVariable Integer idProducto, RedirectAttributes attribute) {
-		productoService.delete(idProducto);
-		attribute.addFlashAttribute("eliminar", "Eliminado con éxito!");
-		return "redirect:/producto";
+	    Optional<Producto> optionalProducto = productoService.get(idProducto);
+	    if (optionalProducto.isPresent()) {
+	        Producto producto = optionalProducto.get();
+	        List<Imagen> imagenes = producto.getImagenesProducto();
+	        
+	        // Ruta absoluta donde se guardan las imágenes
+	        String rutaAbsoluta = "D://Ecommerce//imagenesProducto";
+	        
+	        // Eliminar físicamente las imágenes
+	        for (Imagen imagen : imagenes) {
+	            try {
+	                Path rutaImagen = Paths.get(rutaAbsoluta + "//" + imagen.getArchivoImagen());
+	                Files.deleteIfExists(rutaImagen);
+	            } catch (IOException e) {
+	                e.printStackTrace();
+	            }
+	        }
+	        
+	        productoService.delete(idProducto);
+	        attribute.addFlashAttribute("eliminar", "Eliminado con éxito!");
+	    }
+	    
+	    return "redirect:/producto";
 	}
+
 	
 }
